+++ src/client-cohttp-lwt/opentelemetry_client_cohttp_lwt.ml
@@ -557,7 +557,7 @@ end)
     }
 end
 
-let setup_ ?(stop = Atomic.make false) ~(config : Config.t) () =
+let create_backend ?(stop = Atomic.make false) ?(config = Config.make ()) () =
   debug_ := config.debug;
 
   if config.url <> get_url () then set_url config.url;
@@ -571,12 +571,17 @@ let setup_ ?(stop = Atomic.make false) ~(config : Config.t) () =
       end)
       ()
   in
-  Opentelemetry.Collector.set_backend (module B);
+  (module B : OT.Collector.BACKEND)
+
+let setup_ ?stop ?config () =
+  let backend = create_backend ?stop ?config () in
+  let (module B : OT.Collector.BACKEND) = backend in
+  OT.Collector.set_backend backend;
   B.cleanup
 
-let setup ?stop ?(config = Config.make ()) ?(enable = true) () =
+let setup ?stop ?config ?(enable = true) () =
   if enable then (
-    let cleanup = setup_ ?stop ~config () in
+    let cleanup = setup_ ?stop ?config () in
     at_exit cleanup
   )