+++ src/client-cohttp-lwt/opentelemetry_client_cohttp_lwt.mli
@@ -18,6 +18,12 @@ val set_headers : (string * string) list -> unit
 
 module Config = Config
 
+val create_backend :
+  ?stop:bool Atomic.t ->
+  ?config:Config.t ->
+  unit ->
+  (module Opentelemetry.Collector.BACKEND)
+
 val setup :
   ?stop:bool Atomic.t -> ?config:Config.t -> ?enable:bool -> unit -> unit
 (** Setup endpoint. This modifies {!Opentelemetry.Collector.backend}.