+++ src/opentelemetry.ml
@@ -1,7 +1,5 @@
 (** Opentelemetry types and instrumentation *)
 
-module Thread_local = Thread_local
-
 module Lock = Lock
 (** Global lock. *)
 
@@ -149,6 +147,69 @@ module Collector = struct
 
   type backend = (module BACKEND)
 
+  module Noop_backend : BACKEND = struct
+    let noop_sender _ ~ret = ret ()
+
+    let send_trace : Trace.resource_spans list sender = { send = noop_sender }
+
+    let send_metrics : Metrics.resource_metrics list sender =
+      { send = noop_sender }
+
+    let send_logs : Logs.resource_logs list sender = { send = noop_sender }
+
+    let signal_emit_gc_metrics () = ()
+
+    let tick () = ()
+
+    let set_on_tick_callbacks _cbs = ()
+
+    let cleanup () = ()
+  end
+
+  module Debug_backend (B : BACKEND) : BACKEND = struct
+    open Proto
+
+    let send_trace : Trace.resource_spans list sender =
+      {
+        send =
+          (fun l ~ret ->
+            Format.eprintf "SPANS: %a@."
+              (Format.pp_print_list Trace.pp_resource_spans)
+              l;
+            B.send_trace.send l ~ret);
+      }
+
+    let send_metrics : Metrics.resource_metrics list sender =
+      {
+        send =
+          (fun l ~ret ->
+            Format.eprintf "METRICS: %a@."
+              (Format.pp_print_list Metrics.pp_resource_metrics)
+              l;
+            B.send_metrics.send l ~ret);
+      }
+
+    let send_logs : Logs.resource_logs list sender =
+      {
+        send =
+          (fun l ~ret ->
+            Format.eprintf "LOGS: %a@."
+              (Format.pp_print_list Logs.pp_resource_logs)
+              l;
+            B.send_logs.send l ~ret);
+      }
+
+    let signal_emit_gc_metrics () = B.signal_emit_gc_metrics ()
+
+    let tick () = B.tick ()
+
+    let set_on_tick_callbacks cbs = B.set_on_tick_callbacks cbs
+
+    let cleanup () = B.cleanup ()
+  end
+
+  let debug_backend : backend = (module Debug_backend (Noop_backend))
+
   (* hidden *)
   open struct
     let on_tick_cbs_ = ref []
@@ -195,6 +256,14 @@ module Collector = struct
     match !backend with
     | None -> ()
     | Some (module B) -> B.tick ()
+
+  let with_setup_debug_backend b ?(enable = true) () f =
+    let (module B : BACKEND) = b in
+    if enable then (
+      set_backend b;
+      Fun.protect ~finally:B.cleanup f
+    ) else
+      f ()
 end
 
 module Util_ = struct
@@ -366,6 +435,7 @@ type value =
   [ `Int of int
   | `String of string
   | `Bool of bool
+  | `Float of float
   | `None
   ]
 
@@ -379,6 +449,7 @@ let _conv_value =
   | `Int i -> Some (Int_value (Int64.of_int i))
   | `String s -> Some (String_value s)
   | `Bool b -> Some (Bool_value b)
+  | `Float f -> Some (Double_value f)
   | `None -> None
 
 (**/**)
@@ -521,26 +592,23 @@ module Scope = struct
     if Collector.has_backend () then
       scope.attrs <- List.rev_append (attrs ()) scope.attrs
 
-  (**/**)
+  (** The opaque key necessary to access/set the ambient scope with
+      {!Ambient_context}. *)
+  let ambient_scope_key : t Ambient_context.key = Ambient_context.create_key ()
 
-  let _global_scope : t Thread_local.t = Thread_local.create ()
-
-  (**/**)
-
-  (** Obtain current scope from thread-local storage, if available *)
-  let get_surrounding ?scope () : t option =
+  (** Obtain current scope from {!Ambient_context}, if available. *)
+  let get_ambient_scope ?scope () : t option =
     match scope with
     | Some _ -> scope
-    | None -> Thread_local.get _global_scope
+    | None -> Ambient_context.get ambient_scope_key
 
-  (** [with_scope sc f] calls [f()] in a context where [sc] is the
-      (thread)-local scope, then reverts to the previous local scope, if any. *)
-  let[@inline] with_scope (sc : t) (f : unit -> 'a) : 'a =
-    Thread_local.with_ _global_scope sc (fun _ -> f ())
-end
+  (** [with_ambient_scope sc thunk] calls [thunk()] in a context where [sc] is
+      the (thread|continuation)-local scope, then reverts to the previous local
+      scope, if any.
 
-open struct
-  let get_surrounding_scope = Scope.get_surrounding
+      @see <https://github.com/ELLIOTTCABLE/ocaml-ambient-context> ambient-context docs *)
+  let[@inline] with_ambient_scope (sc : t) (f : unit -> 'a) : 'a =
+    Ambient_context.with_binding ambient_scope_key sc (fun _ -> f ())
 end
 
 (** Span Link
@@ -614,7 +682,13 @@ module Span : sig
   val id : t -> Span_id.t
 
   type key_value =
-    string * [ `Int of int | `String of string | `Bool of bool | `None ]
+    string
+    * [ `Int of int
+      | `String of string
+      | `Bool of bool
+      | `Float of float
+      | `None
+      ]
 
   val create :
     ?kind:kind ->
@@ -651,7 +725,13 @@ end = struct
     | Span_kind_consumer
 
   type key_value =
-    string * [ `Int of int | `String of string | `Bool of bool | `None ]
+    string
+    * [ `Int of int
+      | `String of string
+      | `Bool of bool
+      | `Float of float
+      | `None
+      ]
 
   type nonrec status_code = status_status_code =
     | Status_code_unset
@@ -719,22 +799,14 @@ module Trace = struct
 
   let add_attrs = Scope.add_attrs [@@deprecated "use Scope.add_attrs"]
 
-  (** Sync span guard.
-
-      @param force_new_trace_id if true (default false), the span will not use a
-      surrounding context, or [scope], or [trace_id], but will always
-      create a fresh new trace ID.
-
-      {b NOTE} be careful not to call this inside a Gc alarm, as it can
-      cause deadlocks. *)
-  let with_ ?(force_new_trace_id = false) ?trace_state ?service_name
+  let with_' ?(force_new_trace_id = false) ?trace_state ?service_name
       ?(attrs : (string * [< value ]) list = []) ?kind ?trace_id ?parent ?scope
-      ?links name (f : Scope.t -> 'a) : 'a =
+      ?links name cb =
     let scope =
       if force_new_trace_id then
         None
       else
-        get_surrounding_scope ?scope ()
+        Scope.get_ambient_scope ?scope ()
     in
     let trace_id =
       match trace_id, scope with
@@ -753,8 +825,6 @@ module Trace = struct
     let start_time = Timestamp_ns.now_unix_ns () in
     let span_id = Span_id.create () in
     let scope = { trace_id; span_id; events = []; attrs } in
-    (* set global scope in this thread *)
-    Scope.with_scope scope @@ fun () ->
     (* called once we're done, to emit a span *)
     let finally res =
       let status =
@@ -773,10 +843,39 @@ module Trace = struct
       in
       emit ?service_name [ span ]
     in
+    let thunk () =
+      (* set global scope in this thread *)
+      Scope.with_ambient_scope scope @@ fun () -> cb scope
+    in
+    thunk, finally
+
+  (** Sync span guard.
+
+      Notably, this includes {e implicit} scope-tracking: if called without a
+      [~scope] argument (or [~parent]/[~trace_id]), it will check in the
+      {!Ambient_context} for a surrounding environment, and use that as the
+      scope. Similarly, it uses {!Scope.with_ambient_scope} to {e set} a new
+      scope in the ambient context, so that any logically-nested calls to
+      {!with_} will use this span as their parent.
+
+      {b NOTE} be careful not to call this inside a Gc alarm, as it can
+      cause deadlocks.
+
+      @param force_new_trace_id if true (default false), the span will not use a
+      ambient scope, the [~scope] argument, nor [~trace_id], but will instead
+      always create fresh identifiers for this span *)
+
+  let with_ ?force_new_trace_id ?trace_state ?service_name ?attrs ?kind
+      ?trace_id ?parent ?scope ?links name (cb : Scope.t -> 'a) : 'a =
+    let thunk, finally =
+      with_' ?force_new_trace_id ?trace_state ?service_name ?attrs ?kind
+        ?trace_id ?parent ?scope ?links name cb
+    in
+
     try
-      let x = f scope in
+      let rv = thunk () in
       finally (Ok ());
-      x
+      rv
     with e ->
       finally (Error (Printexc.to_string e));
       raise e